Fix: length(...) function: different behaviour from the specification
On the specification of the VCSSL, length(a, d)
should returns the length of the "d"-th dimension
of the array "a" as a scalar value.
The Vnano is a subset of the VCSSL,
so behaviour of length(a, d) function of the Vnano
should be the same with the above.

However, before this commit,
the length(a, d) function had returned an array
storing lengths of all dimensions of the array "a",
and the value of "d" had not been referred.

This incorrect behaviour has been fixed by this commit.