Fix: TerminalWindow: check whether the caller is the event dispatch thread or not
Some applications might connect/disconnect plug-ins
in the event dispatch thread, so init() and dispose()
methods of TerminalWIndow class might be called from
the procedures running on the event dispatch thread.

Other methods also might be invoked on
the event dispatch thread,
so we have added some lines in them for
checking the caller thread and branching.