Fix: File I/O plug-ins: base directory of relative paths
The base directory of relative file-path resolution
should be the same with the directory
in which the running script is.
The path of the base directory is set as
the value of the "MAIN_SCRIPT_DIRECTORY" option.

However, before this commit, implementations of
file I/O plug-ins had not referred the above value,
and the base directory was always set to "."
(the current directory of the process
 on which the scripting engine runs).

The above bug has been fixed by this commit.